bug fix + cambios
Se puede cambiar el monto por express y el monto por empaque.
Arreglado un error relacionado con intentar mostar imagenes

+++ codigo/src/controladores/ServerPetition.java
@@ -101,7 +101,7 @@ public class ServerPetition {
             Pedido.setPorExpress(express);
             return peticion;
         }else if(tipo==TPeticion.MODRECOGER) {
-            int recoger = (Integer)peticion.getDatos();
+            Double recoger = (Double)peticion.getDatos();
             Pedido.setPorRecoger(recoger);
             return peticion;
         }
